implement FCxtrans() as xtrans parallel to FC()
Replaces sundry macros or direct uses of xtrans data. Note that
FCxtrans() is defined slightly differently in various iops, with the
simplest possible definiton per iop. For those iops where FCxtrans()
is always relative to the topleft of roi, use that as a parameter.
Similarly the necessity to +6 the row/col in cases where these may be
>= -6. Also bypass const-correctness hoops in type decl of xtrans
param in a couple cases.

Also, add one more (size_t).