share output buffers
Am suspicious of this commit. It seems to me that as the pointers do
not change, and the different threads writing to their underlying data
do not have overlapping writes, that there is no need to enforce
sharing (or synchronization). Still, have changed this in accord with
overall dt usage, dt dev request, and my inexperience with OpenMP.
Review/explanation of this from those with OpenMP knowledge is most
appreciated!

(Note that had to redeclare these pointers without *const so that they
are allowable in shared() clauses.)