imageop_math: only use an int lookup
No need to have a separate float lookup for weights, as the weights
are always either 1, 0.5, or 0.25, so just can store their reciprocals
as ints.

Also remove a bit of x-trans-ish code, as presumably this'll need a
rewrite to work with x-trans.

And misc. cleanups: No debugging output, and note that don't take
advantage of regular nature of CFA.