vectorscope: WIP
First pass at white point display on CIE 1931 xy, but need to be
mindful of D50/D65 of the histogram profile -- this currently shows
the wrong white point for Pro Photo which is D50.

Add a whitepoint readout to stdout to test this.

Pulled in color math code to see what happens if I do a absolute
colorimetric matrix for histogram profile -- the answer is no
difference.

Other to do notes.