pixelpipe: don't use cache for expanded iops with changed input
This updates the histogram and picker for a scope if an iop prior to
it in the pixelpipe changes. It does this by not using the cache in
this case.

To test: Set up an image with rgb curve and exposure iops
enabled. Expand rgb curve. Toggle on/off the exposure iop. The rgb
curve histogram should update.