pixelpipe: alternate method to skip cache
Make cache depend on pixelpipe timestamp for cases when it isn't
useful (gamma or when prior module needs a UI update).

This still doesn't work for updating picker, but presumably that
should be redrawn disabled in case of a prior module changing?

This loses the chance to decrease weight of hash lines which we know
we won't use.