color picker: make a flag to turn denoising on/off
Denoising was implemented to help colorpicker black point
detection. In other cases, it just uses extra memory and CPU cycles (a
lot, like 5x more, albeit for a very fast operation). And may hide the
true nature of the image texture from the colorpicker.

QUERY: Should denoising only be on for the auto tune and black pickers
in filmic? As it matters less for the white picker? Or keep for the
white picker to keep its results consistent with the auto picker?