color picker: remove NOP colorspace transformation
Before picking, there was code to transform the image colorspace to
the colorspace the picker requested. Remove this code as in all cases
it was a NOP (source colorspace = destination colorspace), and just
adds verbiage to the source. Also, the colorpicker code is now
colorspace aware, so if there were a colorspace conversion needed, it
should happen in the colorpicker code.