print: GTK-izing of print view, WIP, using GObject
This is a kitchen-sink revamp of print view. The main effort is to use
GTK widgets for all the print view elements. Another effort is to use
a MVC-type setup, with a GObject external to print_settings lib or
print view to hold state.

There's still a bunch of placeholder code. There are no overlays
displayed over the layout boxes for resizing, but they can be dragged
to move. The "callouts" for layout box dimensions are not
implemented. And there is a bunch of code in print_settings which is
bypassed and should be deleted. A few long-standing bugs may have been
fixed, and some code been rewritten in a tidier way.

For now I'm putting this effort on hold -- it's been good to
understand what the print view could turn into, but it is just too
invasive a change. My plan is to use lessons learned from this for a
more succinct GTK-ifying effort.