print: a more minimal GTK-ifying revamp
This is an in-progress snapshot of a more minimal GTK-ifying of print
view. The goal is to keep the existing code structure (rather than
refactor extensively). But still represent the different components of
the print view (background, margins, page body, layout boxes +
grid + callouts) as GTK objects. This is to allow for faster & more
flexible display updates. The plan is to remove some custom code which
GTK can do well (e.g. detection of which box is clicked).

This code should basically work, but it is in no way done. There need
to be separate GTK widgets for each layout box, for the grid, and for
the callouts. There also needs to be more tidying/cleanup.

One change from existing code is how new images are activated from the
filmstrip. I believe that there is a bug in the current code by which
the active image on the filmstrip doesn't change as the user clicks
through images. I have changed the code for this to be closer to
develop view, which should fix this.

Known bug: "working . . ." overlay doesn't go away in center area. I
plan to start using thumbnail widgets to display images, which should
bypass this.