print: WIP gtk-ifying, minimal version
There's some flaky bugs, but general progress. The big rewards of this
aren't here yet -- there's a bunch of boilerplate-ish GTK code but
haven't eliminated the extant code which GTK can supplant.

There are a few small bugfixes, and the initial selection box behavior
is nicer, when the box is drawn outside of the page it is responsive
and snaps into the page when possible.