fix center via overlay setup
The "base" widget for the overlay should be added as a container
contents. This widget provides size allocation for the entire
overlay.

This fixes weirdness where the base drawable for center area was drawn
in front of the overlays, which broke print view once darkroom view
had been entered (which drew on the base center view drawable).

Other than that, this shouldn't affect the other views.

This obviates needing a stub expose event for the print view.