update GUI borderless widget when related print settings change
Rather than on every page expose.

This is still a bit hacky, as we really should have a callback/signal
from print view to update this widget, as print view sets the
borderless flag, and we shouldn't have to know print view's criteria
for this. We could even have dt_view_print_settings() have a return
value.

Also, there's a problem that the logic of displaying the GUI checkbox
isn't the same as when borderless mode is actually enabled. The GUI
logic is better, and the CUPS print code should depend upon it. Better
yet the GUI logic could be improved to handle when there are images
which don't extend to the margins.