print: figure out page dimensions in pixels from GUI
We rely in the GtkAspectFrame to set the page pixel dimensions
proportional the to page measurement, then use those dimensions to
size margins and resize layout boxes.

Move around the image loading a bit, as if we load image before the
page widgets are configured, the image layout boxes are never set and
hence the first loaded image isn't drawn.

There's a difference here from prior code, which figured out the page
pixel dimensions in floating point and drew it relying on Cairo to
handle subpixel detail. Now GTK will be rounding the page dimensions
to the nearest pixel.