cursor and window leave fixups
Reset drag motion when leave view, to deal with if that happens
mid-drag. Fixes weirdness when leave while dragging.

Use old fashioned plus cursor, and change to that cursor when it
enters center area. This cursor is more visible, plus it shows a
physical change when click the "new image area" button.