functional drag-and-drop code for widget-based layout boxes
It should work better than the prior code (fewer unnecessary
refreshes, and responsive when move pointer out of a layout box). It
is still weird in that GTK doesn't seem to understand that portions of
the page that aren't layout boxes aren't a drop zone, even though we
let GTK know this. And still are using some hand-rolled code to detect which
widget is receiving the drop, and it would be nice if we could rely on
GTK for this.